microservicioAuth/requirements.txt
fastapi==0.95.2
pydantic==1.10.12
uvicorn==0.22.0
mysql-connector-python==8.1.0
SQLAlchemy==1.4.15
pymysql==1.0.2
PyJWT==2.6.0
cryptography==39.0.1
python-dotenv==1.0.1
python-multipart==0.0.5
passlib==1.7.4
bcrypt==4.2.0
oauthlib==3.2.2
requests-oauthlib==2.0.0
Werkzeug==3.1.3
